feat (refact): change data type from int to Long

+++ src/main/java/portfolio_urless_backend/url/app/shortenUrl/strategies/ShortenUrlStrategy.java
@@ -1,5 +1,5 @@
 package portfolio_urless_backend.url.app.shortenUrl.strategies;
 
 public interface ShortenUrlStrategy {
-    public String run (int n);
+    public String run (Long n);
 }
